Vendor into a single file
